refactor: extract pluggable StreamSource trait from Yahoo WebSocket
- New StreamSource trait + YahooStreamSource reference impl
- run_stream_loop drives any source with auto-reconnect
- subscribe() defaults to Yahoo; subscribe_with_source() is the generic path

Closes #190.
