+++ docs/server/openapi-html/index.html
@@ -23,7 +23,7 @@
   "openapi": "3.0.3",
   "info": {
     "title": "Finance Query v2",
-    "version": "2.6.1",
+    "version": "2.7.0",
     "description": "REST API for fetching financial data from Yahoo Finance, SEC EDGAR, FRED, CoinGecko, and more.\nProvides real-time quotes, historical data, news, options, SEC filings, macroeconomic series, cryptocurrency data, and more.\n\n**Configuration**\n- EDGAR endpoints require `EDGAR_EMAIL` environment variable.\n- FRED endpoints require `FRED_API_KEY` environment variable.\n\n**Data Sources & Attributions**\n- Market data provided by Yahoo Finance. Not affiliated with or endorsed by Yahoo Finance.\n- SEC filing data sourced from the U.S. SEC EDGAR system (https://www.sec.gov). Not affiliated with the SEC.\n- This product uses the FRED\u00ae API but is not endorsed or certified by the Federal Reserve Bank of St. Louis.\n- U.S. Treasury yield data sourced from the U.S. Department of the Treasury (https://home.treasury.gov).\n- Cryptocurrency data provided by CoinGecko (https://www.coingecko.com). Not affiliated with CoinGecko.\n- Fear & Greed Index data provided by alternative.me (https://alternative.me).\n- News feeds sourced from the Federal Reserve, SEC, MarketWatch, Bloomberg, Financial Times, The Guardian, NYT, Investing.com, BEA, ECB, CFPB, WSJ, Fortune, Business Wire, CoinDesk, CoinTelegraph, TechCrunch, and Hacker News. Content remains the property of respective publishers.\n",
     "contact": {
       "name": "Finance Query",
