+++ finance-query-derive/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "finance-query-derive"
-version = "2.6.1"
+version = "2.7.0"
 edition = "2024"
 authors = ["Harvey Tseng <harveytseng2@gmail.com>"]
 description = "Derive macros for finance-query library"
