+++ server/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "finance-query-server"
-version = "2.6.1"
+version = "2.7.0"
 edition = "2024"
 authors = ["Harvey Tseng <harveytseng2@gmail.com>"]
 description = "High-performance financial data API server built on finance-query library"
