+++ server/asyncapi.yaml
@@ -1,7 +1,7 @@
 asyncapi: 2.6.0
 info:
   title: Finance Query Streaming API
-  version: 2.6.1
+  version: 2.7.0
   description: |
     Real-time price streaming from Yahoo Finance via WebSocket.
